Do not parse XML config and do not start the dbus-deamon
